requirements.txt
pyacoustid>=1.3.0
musicbrainzngs>=0.7.1
mutagen>=1.47.0
tkmacos>=1.0.0; sys_platform == 'darwin'
